Improve account-match observability and break circular import smell
- Add ACCOUNT_MATCH_OUTCOMES_TOTAL counter with per-outcome labels
  (exact, fuzzy_accepted, fuzzy_warned, slugify_new, table_missing).
- Emit a warning when auto-accept is requested but the top fuzzy
  candidate has no account_id, instead of silently slugifying.
- Extract normalize_description plus its compiled regex constants into
  src/moneybin/services/_text.py so categorization_service and
  auto_rule_service both import the helper from a leaf module.
